requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.9
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1
django-allauth==0.51.0
django-model-path-converter==0.1.0
django-model-utils==4.3.1
django-modeltranslation==0.18.4
django-redis==5.2.0
django-reversion==5.0.2
django-prometheus
executing==0.10.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.3
ipython==8.4.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prettydiff==0.1.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pylint==2.15.2
pyparsing==3.0.9
python-dotenv==1.0.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.2
stack-data==0.4.0
tomli==2.0.1
tomlkit==0.11.4
traitlets==5.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
